Fix sign of oscillation frequency in plotting routines
Since the form of the phase is exp(i*(k\cdotx - \omega t)), but we calculate
the growth rate leading to a form of the phase exp(\gamma t + ik\cdot x),
then -i\omega = \gamma. This means that real(\gamma) = imag(\omega), but
that imag(\gamma) = -real(\omega). So we need to apply a minus sign when
plotting the oscillation frequency.