Fix sign error in oscillation frequency
Since typical convention is to write the time behavior as exp(-i\omega
t), where increasing time leads to decreasing phase, and since in the
calculations we use the time behavior exp(\gamma t), this means that
\gamma = -i\omega. So Real(\gamma) + i Imag(\gamma) = -i(Real(\omega)
+ i Imag(\omega)). So Real(\gamma) = Imag(\omega), and the growth rate
requires no sign change. But i Imag(\gamma) = -i Real(\omega), meaning
Imag(\gamma) = -Real(\Omega). So we have to make use of a minus sign
to get the correct sign for the oscillation frequency.